msm8974: Align CPU/GPU frequency corners with AC revision
Although msm8974pro appeared later with an improved crystal, it is
possible for earlier msm8974 family to handle the maximum Krait
frequency introduced by their younger brother. Test devices showed
that a new cap can be achieved with an unobtrusive 50mV voltage
difference between a reference frequency and a target one. Thus, it
is beneficial to attain at least 10% performance improvement percpu
with a slight loss in power efficiency.

GPU corner frequency was left behind, as it is not possible to affect
on PLL registers used as the source of it via the kernel space anymore.
They are programmed by a bootloader since msm-copper. However, having
the same frequencies in kernel is pleasantly from aesthetic side.

Change-Id: Ie71e7448812287560670607bce96c9ab3adf811b
Signed-off-by: Alex Saiko <solcmdr@gmail.com>