kernel: Disable maybe-uninitialized warning globally
Several build configurations had already disabled this warning because it
generates a lot of false positives.  But some had not, and it was still
enabled for allmodconfig builds, for example.

Looking at the warnings produced, every single one I looked at was a false
positive, and the warnings are frequent enough (and big enough) that they
can easily hide real problems that you don't notice in the noise generated
by -Wmaybe-uninitialized.

The warning is good in theory, but this is a classic case of a warning
that causes more problems than the warning can solve.

If gcc gets better at avoiding false positives, we may be able to
re-enable this warning.  But as is, we're better off without it, and I
want to be able to see the *real* warnings.

Change-Id: If9f410f39ee63b06561d01b30a1e1e4f72a60e57
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Alex Saiko <solcmdr@gmail.com>